@@ -7,20 +7,25 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@mantine/core": "^3.6.7",
-    "@mantine/hooks": "^3.6.7",
+    "@mantine/carousel": "^5.0.3",
+    "@mantine/core": "^5.0.3",
+    "@mantine/hooks": "^5.0.3",
+    "@mantine/modals": "^5.0.3",
+    "@tanstack/react-query": "^4.0.10",
+    "@tanstack/react-query-devtools": "^4.0.10",
     "axios": "^0.25.0",
+    "embla-carousel-react": "^7.0.0",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.3.1",
-    "react-query": "^3.34.12",
+    "react-query": "^4.0.0",
     "react-router": "^6.2.1",
     "react-router-dom": "^6.2.2"
   },
   "devDependencies": {
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
     "@typescript-eslint/eslint-plugin": "^5.10.2",
     "@typescript-eslint/parser": "^5.10.2",
     "@vitejs/plugin-react": "^1.0.7",
