@@ -1,12 +1,15 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { BrowserRouter as Router } from 'react-router-dom';
 import StyleProvider from './components/StyleProvider/StyleProvider';
 import QueryProvider from './components/QueryProvider/QueryProvider';
 import './styles/index.css';
 import App from './App';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container!);
+
+root.render(
   <React.StrictMode>
     <Router>
       <StyleProvider>
@@ -15,6 +18,5 @@ ReactDOM.render(
         </QueryProvider>
       </StyleProvider>
     </Router>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 );
