@@ -1,14 +1,17 @@
 import { Container, createStyles } from '@mantine/core';
+import Header from './Header/Header';
 
 // this type is shared with the StyleProvider Component
-type ChildrenType = JSX.Element | React.FC | null;
+type ChildrenType = React.ReactNode | null;
 
 const useStyles = createStyles({
   container: {
     display: 'flex',
-    flexGrow: '1',
+    flexDirection: 'column',
+    flexGrow: 1,
     margin: '0',
     maxWidth: '100%',
+    minHeight: '100vh',
     padding: '0'
   }
 });
@@ -19,6 +22,7 @@ function GeneralLayoutContainer(props: { children: ChildrenType | ChildrenType[]
 
   return (
     <Container className={classes.container}>
+      <Header />
       { children }
     </Container>
   );
