@@ -1,14 +1,15 @@
 import React from 'react';
 import { MantineProvider } from '@mantine/core';
+import CustomTheme from './CustomMantineTheme';
 
-type ChildrenType = JSX.Element | React.FC | null;
+type ChildrenType = React.ReactNode | null;
 
 function StyleProvider(props: { children: ChildrenType | ChildrenType[] }) {
   const { children } = props;
 
   return (
-    <MantineProvider theme={{ colorScheme: 'dark' }} withNormalizeCSS withGlobalStyles>
-      {children}
+    <MantineProvider theme={CustomTheme} withNormalizeCSS withGlobalStyles>
+      { children }
     </MantineProvider>
   );
 }
